// backend/src/interfaces/common.ts
// src/interfaces/common.ts
export type IGenericErrorMessage = {
  path: string | number;
  message: string;
};

export type IGenericErrorResponse = {
  statusCode: number;
  message: string;
  errorMessages: IGenericErrorMessage[];
};

export type IGenericResponse<T> = {
  meta?: { page: number; limit: number; total: number };
  data: T;
};